requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
pydantic-settings
aiomqtt
gmqtt
numpy
aioftp
python-multipart
httpx
aiofiles
